requirements.txt
blinker==1.6.2
click==8.1.7
flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
werkzeug==2.3.7
zipp==3.16.2
